chore: move nova test trait to Tests
